+++ CHANGELOG.md
@@ -134,3 +134,4 @@ All notable changes to this project will be documented in this file.
 - add repeat keyword
 - implement new type manager which keeps better track of types and properties
 - fix and improve documentation regarding intrinsics
+- add description to signature help provider